Altered: - Moved AMXRFScan to acquamanCommon.pri from VESPERS... because... you know... it's an AM class.
