+++ source/acquaman/VESPERS/VESPERSXRFScanController.h
@@ -24,7 +24,7 @@ along with Acquaman.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "acquaman/VESPERS/VESPERSXRFScanConfiguration.h"
 #include "acquaman/AMScanController.h"
-#include "dataman/VESPERS/AMXRFScan.h"
+#include "dataman/AMXRFScan.h"
 #include "beamline/AMXRFDetector.h"
 
 class VESPERSXRFScanController : public AMScanController
